Crate minijinja_contrib

MiniJinja-Contrib is a utility crate for MiniJinja that adds support for certain utilities that are too specific for the MiniJinja core. This is usually because they provide functionality that Jinja2 itself does not have.

To add all of these to an environment you can use the add_to_environment function.

use minijinja::Environment;

let mut env = Environment::new();
minijinja_contrib::add_to_environment(&mut env);

Modules

filters

Utility filters.

globals

Globals

pycompat

Implements Python methods for better compatibility.

Functions

add_to_environment

Registers all features of this crate with an Environment.