minijinja_contrib

Function add_to_environment

pub fn add_to_environment(env: &mut Environment<'_>)

Registers all features of this crate with an Environment.

All the filters that are available will be added, same with global functions that exist.

Note: the pycompat support is intentionally not registered with the environment.